requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
langchain
langchain-core
langchain-community
langchain-ollama
langchain-huggingface
transformers
torch
Pillow
pdf2image
python-multipart
redis
python-jose
passlib[bcrypt]
docx
faiss-cpu
streamlit
requests